Easy way how to make my own operating
Hello! 
This file shows an example of how to make my own operating symbol!
Remember when always the operating reference from the original variables.
Therefore, without malloc or your own memory structure system, 
it always copies the original value in stack memory.
So, after finish the operating function, the copy value is destroyed.
The best way is using the address of value or changes the original value after calculating!